+++ beeref/constants.py
@@ -16,5 +16,24 @@
 APPNAME = 'BeeRef'
 APPNAME_FULL = f'{APPNAME} Reference Image Viewer'
 VERSION = '0.1.1'
-WEBSITE = 'https://github.com/rbreu/beeref'
+WEBSITE = 'https://beeref.org'
 COPYRIGHT = 'Copyright © 2021 Rebecca Breu'
+
+COLORS = {
+    # Qt:
+    'Active:Base': (60, 60, 60),
+    'Active:Window': (40, 40, 40),
+    'Active:Button': (40, 40, 40),
+    'Active:Text': (200, 200, 200),
+    'Active:HighlightedText': (255, 255, 255),
+    'Active:WindowText': (200, 200, 200),
+    'Active:ButtonText': (200, 200, 200),
+    'Active:Highlight': (83, 167, 165),
+    'Active:Link': (90, 181, 179),
+    'Disabled:Light': (0, 0, 0),
+    'Disabled:Text': (140, 140, 140),
+
+    # BeeRef specific:
+    'Scene:Selection': (116, 234, 231),
+    'Scene:Canvas': (60, 60, 60),
+}