+++ beeref/gui.py
@@ -37,7 +37,6 @@ class WelcomeOverlay(QtWidgets.QWidget):
         self.setAttribute(Qt.WidgetAttribute.WA_NoSystemBackground)
         self.setAttribute(Qt.WidgetAttribute.WA_TransparentForMouseEvents)
         label = QtWidgets.QLabel(self)
-        label.setStyleSheet("QLabel { color: #cccccc; }")
         label.setText(self.txt)
         label.setAlignment(Qt.AlignmentFlag.AlignVCenter
                            | Qt.AlignmentFlag.AlignCenter)